backend/requirements.txt
asgiref==3.8.1
Django==5.1.5
django-extensions==3.2.3
djangorestframework==3.15.2
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.17.0
sqlparse==0.5.3
tzdata==2024.2
django-cors-headers==4.6.0
django-filter==24.3
elasticsearch==8.17.1
algoliasearch==4.13.0
algoliasearch-django>=4.0,<5.0
python_dotenv==1.0.1